requirements.txt
google-generativeai
pydantic
